reorganized auth module and folder structure
