+++ backend/config/serviceAccountKey.json
@@ -1,4 +1,13 @@
-// To get the service account go to your Firebase Dashboard and 
-// open Settings in the side panel. Then click the Service Accounts
-// tab and click Generate New Private Key, then confirm by clicking Generate Key.
-// This will download a JSON file to your machine. Copy the contents here.
+{
+    "type": "service_account",
+    "project_id": "f2f-auth-3c4db",
+    "private_key_id": "a0f4b0be4102107a5a2cc9c94ab17a08f839599f",
+    "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCxb2Xc8OUFo6IF\nRdnlKSSKRT7YC3ooJ2bv0PghH7p0eGlyHjsAUf8FumOgioRx++oAtJsbYzHi/uYi\nn0pEykScxtTeeZzyPJyyRBcvlauCJ34H0m4ab/hCuDDj7R3rVLjNXfjmSdQwGdyY\nLAmH8EwU35D+EEbmpDsDOJ/NNkF4HzuTdv58YAsoqZyvHDpTjM7WvbBwQbbw+ufM\nh/yK8Abo3FtlFClLuNumWK/whOwUVTpzE3REsZkG3VlcLDVpDWYtqlVhbCU74Oqs\nwWGsIMe+K1GfNPeVsDbRj4ZgmsrSOud+wZboFZgTsBwOnSwL9+h58V1+3RONWxoj\nBvrkq7B3AgMBAAECggEAUU9jxh5MTUvBtRnf8uBqQsTduiJ0Tc9UoCRKeN9x0erW\nI3oP/qMUszrJaengLN3My5r4E1JfK480xfvqDGEGlTOWnAbt8aHuCtEiMOetCz1u\nx0Xat9nE0kpW+TQRiGSwNUhRhbKifpJzYPaZPYbUTqJu7sZlv07WTjEWJfOth84+\nfOVuunARK+3dB3tacejLaZsc3K6wSpNNL9eJZe+jkH2h3+OHud/EsYraI91lgdlt\n1AxbNIn/X/tSo51+oMM4C+t1jc5sz5nJFKHM/KKZ702/1m3aHszEvJu8lcKDO9PJ\nRyMvs9H1q4DYNiupfKLEPIpRxeGhbn7ZNMoi1xXDXQKBgQD5bBGvCfcLuwIwnuoh\nhhxGKIypybrzTEXyAcPXGV0wC2SdZkuE6dKShrefPpyIbk8V/9cSuHGMQ6etnz04\nYw870Ow0xe+RLshAlxj9v0eZpknCOEHhHo6z2uOPzzD2nPoHfw8hG6oLRgejALHJ\nRmWigzQ94LG3V52QuA41OPGbHQKBgQC2HVIo7o+mPIiYVriH9Fq0iJQmE4pqwqa3\n5+dndojH53pFTWb/YdtjEpDKiRb92O4yWVAAWTOvtrHQlpY1MwS3xqEFpcv1PznK\nsKKrbh2V2XQKlRbBpR+mFGMQO+7SbYeDaAynrnX2WESySydGWZGilG4K8w/Le4Dj\n7yBg1l0RowKBgECHF/Ab+OD9Ca0Af/waURhJh1Kp0Qb+tZ3NnRf6BOTYHbim7WS+\nhW7JEZpOPXsCZs1/Oeoq851hJ3FAoDsFTbyg6Wshxj/eRYqITLvhD6oNnklz+n8+\nQ5b+Sx86cCVtKTZlRcm/+XcfL/yYwmYbI1Iu9YYIqpGPRkFZjLtyNvd1AoGBALKJ\nyFRsnNQCoH+I9VwECTDYfTVfDuJfdobxiYF7SdmBa/OpUIKqcySLvPtyruGq3Mrr\nmGzfBrcLSwQNvbDmFYf96TraMjx7AbNeu/uI9uRDzWnag/bQZViyv/PYbehKjQIc\n7dsuRGOAYcjRuO1VPlu/XKrLY/6/U7siYgHN5mw1AoGAfRbjNqLN+m83Fejh/eQP\nTHazns0i6gFy2Je7/7SLFT1RDpYNYujZJBHxqC8az9gSEC2JOHKI8TC1yyL5xUv0\nLsYIGQ3PD7TQ/u8UVSBYjVEZgB44GyLU5U0sYwgnOq8QMdy4IT/zm3W9PhxL5Mv8\nax5hESgJuO8C+ipL9gWlpvY=\n-----END PRIVATE KEY-----\n",
+    "client_email": "firebase-adminsdk-hc9jc@f2f-auth-3c4db.iam.gserviceaccount.com",
+    "client_id": "108361989886012092327",
+    "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+    "token_uri": "https://oauth2.googleapis.com/token",
+    "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+    "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-hc9jc%40f2f-auth-3c4db.iam.gserviceaccount.com"
+  }
+
