+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import { BrowserRouter } from "react-router-dom";
-import Router from "./Router";
+import Router from "./lib/Router";
 
 const App = () => {
   return (
