+++ frontend/src/lib/constants.js
@@ -1,0 +1,9 @@
+export const WINDOW_TYPE = {
+    WEB: 'WEB',
+    MOBILE: 'MOBILE'
+};
+
+export const STATUS_CODE = {
+    SUCESS : "SUCESS",
+    ERROR: "ERROR"
+}
