+++ frontend/src/lib/firebase.js
